🔨 refactor: JoinPage 컴포넌트와 JoinForm 컴포넌트 합치기
